Docstrings, moved MongoCollection.wacth to a new file
