+++ Sources/MongoSwift/MongoClient.swift
@@ -306,7 +306,7 @@ public class MongoClient {
 
      /**
       * Starts a `ChangeStream` on a `MongoClient`. Allows the client to observe all changes in a cluster - excluding
-      * system collections i.g. "config", "local", and "admin" databases.
+      * system collections and the "config", "local", and "admin" databases.
       * - Parameters:
       *   - pipeline: An array of aggregation pipeline stages to apply to the events returned by the change stream.
       *   - options: An optional `ChangeStreamOptions` to use when constructing the change stream.
@@ -333,13 +333,14 @@ public class MongoClient {
 
      /**
       * Starts a `ChangeStream` on a `MongoClient`. Allows the client to observe all changes in a cluster - excluding
-      * system collections i.g. "config", "local", and "admin" databases. Associates the specified `Codable` type `T`
-      * with the `fullDocument` field in the `ChangeStreamDocument` emitted by the returned `ChangeStream`.
+      * system collections and the "config", "local", and "admin" databases. Associates the specified `Codable` type `T`
+      * with the `fullDocument` field in the `ChangeStreamDocument`s emitted by the returned `ChangeStream`.
       * - Parameters:
       *   - pipeline: An array of aggregation pipeline stages to apply to the events returned by the change stream.
       *   - options: An optional `ChangeStreamOptions` to use when constructing the change stream.
       *   - session: An optional `ClientSession` to use with this change stream.
-      *   - withFullDocumentType: The type that the change events emitted from the change stream will be decoded to.
+      *   - withFullDocumentType: The type that the `fullDocument` field of the emitted `ChangeStreamDocument`s will be
+      *                           decoded to.
       * - Returns: A `ChangeStream` on all collections in all databases in a cluster.
       * - Throws:
       *   - `ServerError.commandError` if an error occurs on the server while creating the change stream.
@@ -366,13 +367,14 @@ public class MongoClient {
 
      /**
       * Starts a `ChangeStream` on a `MongoClient`. Allows the client to observe all changes in a cluster - excluding
-      * system collections i.g. "config", "local", and "admin" databases. Associates the specified `Codable` type `T`
+      * system collections and the "config", "local", and "admin" databases. Associates the specified `Codable` type `T`
       * with the returned `ChangeStream`.
       * - Parameters:
       *   - pipeline: An array of aggregation pipeline stages to apply to the events returned by the change stream.
       *   - options: An optional `ChangeStreamOptions` to use when constructing the change stream.
       *   - session: An optional `ClientSession` to use with this change stream.
-      *   - withEventType: The type that the entire change stream response will be decoded to.
+      *   - withEventType: The type that the entire change stream response will be decoded to and that will be returned
+      *                    when iterating through the change stream.
       * - Returns: A `ChangeStream` on all collections in all databases in a cluster.
       * - Throws:
       *   - `ServerError.commandError` if an error occurs on the server while creating the change stream.
